sync vscode-mysql files
