+++ vscode-mysql/config/my.cnf
@@ -1,0 +1,39 @@
+# Copyright (c) 2017, Oracle and/or its affiliates. All rights reserved.
+#
+# This program is free software; you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation; version 2 of the License.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program; if not, write to the Free Software
+# Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301 USA
+
+#
+# The MySQL  Server configuration file.
+#
+# For explanations see
+# http://dev.mysql.com/doc/mysql/en/server-system-variables.html
+
+[mysqld]
+pid-file        = /var/run/mysqld/mysqld.pid
+socket          = /var/run/mysqld/mysqld.sock
+datadir         = /var/lib/mysql
+user            = nobody
+
+# Allow file import and exports
+secure-file-priv= ""
+
+# Disable some logging files for DB recovery purposes, since our use case won't have heavy concurrent writes, and we want to reduce storage if possible
+disable_log_bin
+innodb-doublewrite=OFF
+
+# Set disk space limit to temp table file, which grows until server restart
+innodb_temp_data_file_path=ibtmp1:12M:autoextend:max:5G
+
+# Custom config should go here
+!includedir /etc/mysql/conf.d/
