+++ vscode-mysql/nginx-files/.hidden
@@ -1,0 +1,2 @@
+.hidden
+.readonly
