+++ vscode-mysql/nginx-files/.readonly
@@ -1,0 +1,1 @@
+reverse-proxy.conf
