+++ vscode-mysql/nginx-files/reverse-proxy.conf
@@ -1,0 +1,61 @@
+server {
+    listen 8080;
+    listen [::]:8080;
+
+    ###
+    # DO NOT MODIFY ANYTHING ABOVE THIS LINE
+    ###
+
+    # helpful comments to let instructors know what everything does
+
+    location /serve/ {
+        proxy_pass http://localhost:3000/;
+    }
+
+    location /static/js/ {
+        proxy_pass http://localhost:3000;
+        proxy_http_version 1.1;
+        proxy_set_header Upgrade $http_upgrade;
+        proxy_set_header Connection 'upgrade';
+        proxy_set_header Host $host;
+        proxy_cache_bypass $http_upgrade;
+    }
+
+    location /sockjs-node/ {
+        proxy_pass http://localhost:3000;
+        proxy_http_version 1.1;
+        proxy_set_header Upgrade $http_upgrade;
+        proxy_set_header Connection 'upgrade';
+        proxy_set_header Host $host;
+        proxy_cache_bypass $http_upgrade;
+    }
+    
+    ###
+    # DO NOT MODIFY ANYTHING BELOW THIS LINE
+    ###
+
+    location / {
+        proxy_pass http://localhost:7000; # This port should be the port on which code-server is running
+          proxy_set_header  Host $host;
+        proxy_set_header  X-Real-IP $remote_addr;
+        proxy_set_header  X-Forwarded-Proto https;
+        proxy_set_header  X-Forwarded-For $remote_addr;
+        proxy_set_header  X-Forwarded-Host $remote_addr;
+        proxy_http_version 1.1;
+        proxy_set_header Upgrade $http_upgrade;
+        proxy_set_header Connection "upgrade";
+    }
+
+    # Failsafe to try to prevent reverse-proxy configuration that would break lab launch
+    location = /static/ {
+        proxy_pass http://localhost:7000; # This port should be the port on which code-server is running
+          proxy_set_header  Host $host;
+        proxy_set_header  X-Real-IP $remote_addr;
+        proxy_set_header  X-Forwarded-Proto https;
+        proxy_set_header  X-Forwarded-For $remote_addr;
+        proxy_set_header  X-Forwarded-Host $remote_addr;
+        proxy_http_version 1.1;
+        proxy_set_header Upgrade $http_upgrade;
+        proxy_set_header Connection "upgrade";
+    }
+}
