+++ vscode-mysql/supervisord.conf
@@ -1,0 +1,39 @@
+[unix_http_server]
+file=/var/run/supervisor.sock
+chmod=0770
+chown=nobody:nogroup
+
+[supervisord]
+logfile=/var/log/supervisor/supervisord.log
+pidfile=/var/run/supervisord.pid
+nodaemon=true
+
+[rpcinterface:supervisor]
+supervisor.rpcinterface_factory=supervisor.rpcinterface:make_main_rpcinterface
+
+[supervisorctl]
+serverurl=unix:///var/run/supervisor.sock
+
+[inet_http_server]
+port = 9001
+
+[include]
+files=/etc/supervisor/conf.d/*.conf
+
+[program:nginx]
+command=nginx -g 'daemon off;'
+process_name=%(program_name)s
+priority=1000
+autostart=true
+startretries=3
+autorestart=true
+stdout_logfile=/var/log/supervisor/nginx-stdout.log
+stderr_logfile=/var/log/supervisor/nginx-stderr.log
+user=root
+
+[program:code-server]
+command=code-server --config /tmp/config/code-server/config.yaml --port 7000
+stdout_logfile=/var/log/supervisor/code-server-stdout.log
+stderr_logfile=/var/log/supervisor/code-server-stderr.log
+user=coder
+environment=HOME="/home/coder"
