--- vscode-mysql/v1/config/conf.d/docker.cnf
@@ -1,3 +1,0 @@
-[mysqld]
-skip-host-cache
-skip-name-resolve
