--- vscode-mysql/v1/manifest.json
@@ -1,15 +1,0 @@
-{
-  "name": "base/vscode-mysql-image-v1.0",
-  "httpPort": 8080,
-  "mounts": [
-    {
-      "path": "/home/coder/project"
-    },
-    {
-      "path": "/home/coder/coursera"
-    },
-    {
-      "path": "/var/lib/mysql"
-    }
-  ]
-}
