--- vscode-mysql/v1/vscode-default/config.yaml
@@ -1,4 +1,0 @@
-bind-addr: 0.0.0.0:8080
-auth: none
-cert: false
-extensions-dir: /home/coder/.local/share/code-server/extensions
