--- vscode-mysql/v1/vscode-default/index.html
@@ -1,1 +1,0 @@
-
