--- vscode-mysql/v2/vscode-default/index.html
@@ -1,1 +1,0 @@
-
