+++ vscode-mysql/vscode-default/index.html
@@ -1,0 +1,1 @@
+
